refactor: switch enum to union type
