+++ packages/core/src/execution/dispatch/executeLibraryDeploy.ts
@@ -7,7 +7,6 @@ import type {
 
 import { ContractFactory, ethers } from "ethers";
 
-import { VertexResultEnum } from "../../types/graph";
 import { collectLibrariesAndLink } from "../../utils/collectLibrariesAndLink";
 
 import { resolveFrom, toAddress } from "./utils";
@@ -35,7 +34,7 @@ export async function executeLibraryDeploy(
     });
   } catch (err) {
     return {
-      _kind: VertexResultEnum.FAILURE,
+      _kind: "failure",
       failure: err as any,
     };
   }
@@ -45,12 +44,12 @@ export async function executeLibraryDeploy(
     receipt = await services.transactions.wait(txHash);
   } catch {
     return {
-      _kind: VertexResultEnum.HOLD,
+      _kind: "hold",
     };
   }
 
   return {
-    _kind: VertexResultEnum.SUCCESS,
+    _kind: "success",
     result: {
       name: artifact.contractName,
       abi: artifact.abi,
