+++ packages/core/src/helpers.ts
@@ -3,4 +3,3 @@ export { createServices } from "./services/createServices";
 export { serializeReplacer } from "./utils/serialize";
 export { TransactionsService } from "./services/TransactionsService";
 export { ContractsService } from "./services/ContractsService";
-export { VertexResultEnum } from "./types/graph";
