+++ packages/core/src/services/AccountsService.ts
@@ -1,5 +1,5 @@
 import type { Providers } from "../types/providers";
-import type { IAccountsService } from "./services";
+import type { IAccountsService } from "../types/services";
 import type { ethers } from "ethers";
 
 export class AccountsService implements IAccountsService {
