+++ packages/core/src/services/ArtifactsService.ts
@@ -1,7 +1,7 @@
 import type { Artifact } from "../types/hardhat";
 import type { Providers } from "../types/providers";
 
-import { IArtifactsService } from "./services";
+import { IArtifactsService } from "../types/services";
 
 export class ArtifactsService implements IArtifactsService {
   constructor(private readonly _providers: Providers) {}
