+++ packages/core/src/services/TransactionsService.ts
@@ -1,5 +1,5 @@
 import type { Providers } from "../types/providers";
-import type { ITransactionsService } from "./services";
+import type { ITransactionsService } from "../types/services";
 
 import { ethers } from "ethers";
