+++ packages/core/src/validation/dispatch/validateArtifactLibrary.ts
@@ -1,12 +1,12 @@
-import { ethers } from "ethers";
-
-import { ArtifactLibraryDeploymentVertex } from "../../types/deploymentGraph";
-import { VertexResultEnum } from "../../types/graph";
-import {
+import type { ArtifactLibraryDeploymentVertex } from "../../types/deploymentGraph";
+import type {
   ValidationDispatchContext,
   ValidationResultsAccumulator,
   ValidationVertexVisitResult,
 } from "../../types/validation";
+
+import { ethers } from "ethers";
+
 import { isArtifact } from "../../utils/guards";
 
 import { buildValidationError, validateBytesForArtifact } from "./helpers";
@@ -58,7 +58,7 @@ export async function validateArtifactLibrary(
   }
 
   return {
-    _kind: VertexResultEnum.SUCCESS,
+    _kind: "success",
     result: undefined,
   };
 }
